% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{checktrimfiles}
\alias{checktrimfiles}
\title{Check files after filterAndTrim}
\source{
\href{../R/computation.R#L83}{computation.R}
}
\usage{
checktrimfiles(A, filt.dir, trimlist)
}
\arguments{
\item{A}{mapping data.frame}

\item{filt.dir}{filtered data directory}

\item{trimlist}{list of vectors, R1 and R2 of trimmed files}
}
\value{
list of \code{A}, \code{trimr1}, and maybe \code{trimr2} with missing files/rows removed.
Stops on error if no trimmed files exist.
}
\description{
Check files after filterAndTrim
}
