% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{comb}
\alias{comb}
\title{Combine function foreach loop}
\usage{
comb(x, ...)
}
\arguments{
\item{x}{list of n items to append onto}

\item{...}{individual lists of n items to append onto n lists in x}
}
\value{
list of lists named according to names of list from single iteration.
}
\description{
For a \code{\link[foreach]{foreach}} loop that returns a list of n items for
each iteration, this function combines all iterations into n different lists - one for each
item returned.
}
\details{
Use this as the value of the `.combine` parameter in foreach.  You will
need to also initiate the list of n lists for the `.init` foreach parameter.
}
\examples{

\dontrun{
  `\%do\%` <- foreach::`\%do\%`
  oper <- foreach::foreach(i=1:5, .combine='comb', .multicombine=TRUE,
  .init=list(list(), list())) \%do\% {
    list(i*2, i*3)
  }

  oper[[1]]
}



}
