% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{dada2compute}
\alias{dada2compute}
\alias{trycomputewrapper}
\title{run dada2 pipeline}
\source{
\href{../R/computation.R#L366}{dada2compute: computation.R}

[trycomputewrapper: computation.R](../R/computation.R#L635)
}
\usage{
dada2compute(
  datadir,
  outdir,
  mapfile,
  refdb = getOption("dparams")$refdb,
  refdb_species = getOption("dparams")$refdb_species,
  nthread = TRUE,
  chimera = FALSE,
  trimLeft = getOption("dparams")$trimLeft,
  trimOverhang = getOption("dparams")$trimOverhang,
  data_type = "PE",
  maxEE = getOption("dparams")$maxEE,
  truncQ = getOption("dparams")$truncQ,
  minBoot = getOption("dparams")$minBoot,
  truncLen = getOption("dparams")$truncLen,
  maxMismatch = getOption("dparams")$maxMismatch,
  justConcatenate = getOption("dparams")$justConcatenate,
  taxmethod = getOption("dparams")$taxmethod,
  band_size = NULL,
  homopolymer_gap_penalty = NULL
)

trycomputewrapper(datadir, outdir, mapfile, logfilename = "logfile.txt", ...)
}
\arguments{
\item{datadir}{Directory containing FASTQ files.}

\item{outdir}{Output directory.}

\item{mapfile}{Mapping filename - needs to be tab separated and have, at minimum,
SampleID, ForwardFastqFile, and ReverseFastqFile columns.}

\item{refdb}{Path to the reference database.}

\item{refdb_species}{Path to the species reference database.}

\item{nthread}{Number of processors for parallel steps.}

\item{chimera}{Should chimeric sequences be removed?  If primers are not}

\item{trimLeft}{(Optional).  In \link[dada2:filterAndTrim]{dada2::filterAndTrim}, the number of nucleotides to remove
from the start of each read, forward and reverse.  The values should be chosen based on
the lengths of primers used for sequencing. (Whole number vector of length 2, for forward
and reverse).}

\item{trimOverhang}{(Optional). After merging paired end reads, trim sequence which overhangs
the start of each read. If amplicons are shorter than read length, e.g. 16S V4 region, we
suggest setting this to True.  (Logical. Default: False).}

\item{data_type}{'SE' or 'PE'}

\item{maxEE}{Integer. Remove reads with less than maxEE expected errors in \link[dada2:filterAndTrim]{dada2::filterAndTrim}}

\item{truncQ}{Integer. Truncate reads at first qual below truncQ in \link[dada2:filterAndTrim]{dada2::filterAndTrim}}

\item{minBoot}{Integer.  Bootstrap value for \link[dada2:assignTaxonomy]{dada2::assignTaxonomy}}

\item{truncLen}{Integer.  Truncation length for reads in \link[dada2:filterAndTrim]{dada2::filterAndTrim}.  Can be vector of length 1 or 2}

\item{maxMismatch}{Integer. Maximum number of allowed mismatches in \link[dada2:mergePairs]{dada2::mergePairs}}

\item{justConcatenate}{Logical. Should PE reads be concatenated instead of overlapped and merged.}

\item{taxmethod}{method for taxonomic assignment.  either idtaxa or rdp.}

\item{band_size}{Integer.  Band size for \link[dada2:dada]{dada2::dada}, \link[dada2:setDadaOpt]{dada2::setDadaOpt}}

\item{homopolymer_gap_penalty}{Integer. for \link[dada2:dada]{dada2::dada} & \link[dada2:setDadaOpt]{dada2::setDadaOpt}. cost of gaps in homopolymer
regions (>=3 repeated bases)}

\item{logfilename}{Log file name (full path).}

\item{...}{parameters to pass to \code{dada2compute}}
}
\value{
\code{trycomputewrapper} returns 0 if \code{dada2compute} succeeds and stops on error.
In rpy2, it will throw \code{rpy2.rinterface.RRuntimeError}.
}
\description{
\code{dada2compute} will run the whole pipeline from within R.
See  \code{\link{.onLoad}} examples for argument default values
or \code{getOption('dparams')}.

\code{trycomputewrapper} is a wrapper for \code{dada2compute} to be used with
rpy2 to run the pipeline in python.  It sets up the global R options and the output to the
logfile before running \code{dada2compute}.
}
