% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{decipher_assign}
\alias{decipher_assign}
\title{Assign taxonomy using DECIPHER}
\usage{
decipher_assign(refdb, seqtab, nthread)
}
\arguments{
\item{refdb}{reference database .RData file}

\item{seqtab}{sequence table made by DADA2}

\item{nthread}{number of processors to use}
}
\value{
taxa table in DADA2 taxa format
}
\description{
Assign taxonomy using DECIPHER
}
