% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{getN}
\alias{getN}
\title{Get number of reads}
\usage{
getN(x)
}
\arguments{
\item{x}{dada2 object}
}
\value{
Integer vector
}
\description{
get number of reads for each sample from a dada2 object (such as output of
\code{\link[dada2]{mergePairs}}).
}
