% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomformat.R
\name{make_biom}
\alias{make_biom}
\title{Create a \link[biomformat]{biom-class}
from \code{\link{matrix-class}}
or \code{\link{data.frame}}.}
\source{
[make_biom in biomformat.R](../R/biomformat.R#L53)
}
\usage{
make_biom(
  data,
  sample_metadata = NULL,
  observation_metadata = NULL,
  id = NULL,
  matrix_element_type = "int",
  qiime_format = TRUE
)
}
\arguments{
\item{data}{(Required).
\code{\link{matrix-class}} or \code{\link{data.frame}}.
A contingency table.
Observations / features / OTUs / species are rows,
samples / sites / libraries are columns.}

\item{sample_metadata}{(Optional).
A \code{\link{matrix-class}} or \code{\link{data.frame}}
with the number of rows equal to the number of samples in \code{data}.
Sample covariates associated with the count data.
This should look like the table returned by
\code{\link[biomformat]{sample_metadata}} on a valid instance
of the \code{\link[biomformat]{biom-class}}.}

\item{observation_metadata}{(Optional).
A \code{\link{matrix-class}} or \code{\link{data.frame}}
with the number of rows equal to the number of
features / species / OTUs / genes in \code{data}.
This should look like the table returned by
\code{\link[biomformat]{observation_metadata}} on a valid instance
of the \code{\link[biomformat]{biom-class}}.}

\item{id}{(Optional). Character string. Identifier for the project.}

\item{matrix_element_type}{(Optional). Character string. Either 'int' or 'float'}

\item{qiime_format}{(Optional). Logical. biom-format requires that observation
metadata be key, value pairs (or group,
 dataset for hd5).  For QIIME, there is only one pair with key be set to "taxonomy,"
 and the value must be the entire taxonomy table.  If FALSE, each column of observation
 metadata will be a separate key (to be consistent with sample metadata).}
}
\value{
An object of \code{\link[biomformat]{biom-class}}.
}
\description{
This function creates a valid instance of the \code{\link[biomformat]{biom-class}}
from standard base-R objects like
\code{\link{matrix-class}} or \code{\link{data.frame}}.
The object returned by this function is appropriate for writing to
a \code{.biom} file using the \code{\link{write_biom}} function.
The sparse biom-format is not (yet) supported.
}
\note{
This code is forked from version 1.5 of the \href{https://github.com/joey711/biomformat-oldfork}{biomformat R library}.
}
