% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{.onLoad}
\alias{.onLoad}
\title{global constants}
\source{
[onLoad:computation.R](../R/computation.R#L43)
}
\usage{
.onLoad(libname, pkgname)
}
\description{
Set global constants to programmatically write documentation.  See [user doc](user_doc.md)
for more complete descriptions.  See examples for the actual values.
}
\note{
The dparams option is set to be a list of the individual values as follows:
\describe{
  \item{maxEE}{parameter for dada2::filterAndTrim}
  \item{truncQ}{parameter for dada2::filterAndTrim}
  \item{trimLeft}{parameter for dada2::filterAndTrim}
  \item{nbases}{parameter for dada2::learnErrors}
  \item{minOverlap}{parameter for dada2::mergePairs}
  \item{maxMismatch}{parameter for dada2::mergePairs}
  \item{trimOverhang}{parameter for dada2::mergePairs}
  \item{justConcatenate}{parameter for dada2::mergePairs}
  \item{outputfasta}{filename for sequence variant FASTA file}
  \item{minBoot}{parameter for dada2::assignTaxonomy}
  \item{biomfile}{filename for biom file based on output of dada2::assignTaxonomy}
  \item{speciesbiomfile}{filename for biom file based on output of dada2::addSpecies}
  \item{otutable}{filename for tab delimited otu table based on output of dada2::addSpecies}
  \item{biomsummary}{filename for text file containing summary of OTU table}
  \item{refdb}{database filename}
  \item{refdb_species}{species database filename}
  \item{min_seq_length}{minimum length of denoised sequences to be used for taxonomic assignment.}
  \item{taxmethod}{method of taxonomic assignment}
  \item{taxtable}{filename for taxonomy table output}
}
}
\examples{
getOption("dparams")

}
