% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{run_cmd}
\alias{run_cmd}
\title{wrap command in tryCatch}
\source{
[computation.R](../R/computation.R#L113)
}
\usage{
run_cmd(cmd, step = NULL, bline = 0, aline = 0, w2e = NA, log = T)
}
\arguments{
\item{cmd}{Character. Command string.}

\item{step}{(Optional). Step name to pass to error.  Default will use cmd.}

\item{bline}{Number of blank lines to precede log output; parameter for \code{\link{logoutput}}}

\item{aline}{Number of blank lines to follow log output; parameter for \code{\link{logoutput}}}

\item{w2e}{warning message to escalate to error.}

\item{log}{log command to file}
}
\value{
cmd will be evaluated in the parent environment, so return values in cmd will be there.
}
\description{
(optionally) log cmd to output and evaluate cmd in the parent environment, catching
and parsing errors.
}
