% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomformat.R
\name{write_biom}
\alias{write_biom}
\title{Write a biom-format v1 file, returning a \link[biomformat]{biom-class}.}
\source{
[write_biom in biomformat.R](../R/biomformat.R#L123)
}
\usage{
write_biom(biom, biom_file, pretty = FALSE)
}
\arguments{
\item{biom}{(Required). A biom object.}

\item{biom_file}{(Required). A character string indicating the
file location of the biom formatted file. This is a JSON formatted file
specific to biological datasets.
The format is formally defined at
\href{http://biom-format.org/documentation/biom_format.html}{the biom-format definition}}

\item{pretty}{logical; Should biom output be pretty printed?}
}
\value{
Nothing. The first argument, \code{x}, is written to a file.
}
\description{
Write a biom-format v1 file, returning a \link[biomformat]{biom-class}.
}
\note{
This code is forked from version 1.5 of the \href{https://github.com/joey711/biomformat-oldfork}{biomformat R library}.
}
